+++ src/Products/MeetingCharleroi/tests/testWFAdaptations.py
@@ -109,7 +109,7 @@ class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
         self.assertFalse('backToProposedFromPrevalidated' in itemWF.transitions)
         self.assertTrue('backToItemCreatedFromPrevalidated' in itemWF.transitions)
 
-        # together with 'pre_validation' and 'harleroi_add_refadmin', it is ok
+        # together with 'pre_validation' and 'charleroi_add_refadmin', it is ok
         cfg.setWorkflowAdaptations(('pre_validation',
                                     'charleroi_add_refadmin',
                                     'charleroi_return_to_any_state_when_prevalidated'))
@@ -117,6 +117,27 @@ class testWFAdaptations(MeetingCharleroiTestCase, mctwfa):
         itemWF = self.wfTool.getWorkflowsFor(cfg.getItemTypeName())[0]
         self.assertTrue('backToProposedFromPrevalidated' in itemWF.transitions)
         self.assertTrue('backToItemCreatedFromPrevalidated' in itemWF.transitions)
+
+        # test that transitions do the work
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem')
+
+        # back to proposed to servicehead
+        self.do(item, 'propose')
+        self.do(item, 'proposeToRefAdmin')
+        self.do(item, 'prevalidate')
+        # MeetingMember can not trigger the transitions
+        self.changeUser('pmCreator1')
+        self.assertEqual(self.transitions(item), [])
+        self.changeUser('pmManager')
+        self.do(item, 'backToProposedFromPrevalidated')
+        self.assertEqual(item.queryState(), 'proposed')
+
+        # back to itemcreated
+        self.do(item, 'proposeToRefAdmin')
+        self.do(item, 'prevalidate')
+        self.do(item, 'backToItemCreatedFromPrevalidated')
+        self.assertEqual(item.queryState(), 'itemcreated')
 
     def _waiting_advices_with_prevalidation_active(self):
         '''Enable WFAdaptation 'charleroi_add_refadmin' before executing test.'''