+++ CHANGES.rst
@@ -6,6 +6,7 @@ Products.MeetingCharleroi Changelog
 
 - Be defensive when using getProperty on a member object, do not fail if member is None
 - Category 'indeterminee' can not be used on MeetingItemCollege if not to send to 'meting-config-council'
+- Added possibility to send and item that is 'prevalidated' back to 'proposed' and 'itemcreated'
 
 4.0 (2017-08-22)
 ----------------